Refactor plugin to capture tracked events
